requirements.txt
# Common dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiohttp>=3.9.0
yt-dlp>=2024.0.0

# Windows dependencies (only installed on Windows) - pre-built wheels
winrt-runtime>=2.0.0; sys_platform == "win32"
winrt-Windows.Foundation>=2.0.0; sys_platform == "win32"
winrt-Windows.Foundation.Collections>=2.0.0; sys_platform == "win32"
winrt-Windows.ApplicationModel>=2.0.0; sys_platform == "win32"
winrt-Windows.Media.Control>=2.0.0; sys_platform == "win32"
winrt-Windows.Storage.Streams>=2.0.0; sys_platform == "win32"
winrt-Windows.UI.Notifications>=2.0.0; sys_platform == "win32"
winrt-Windows.UI.Notifications.Management>=2.0.0; sys_platform == "win32"

# Linux dependencies (only installed on Linux)
dbus-python>=1.3.2; sys_platform == "linux"
PyGObject>=3.46.0; sys_platform == "linux"

# Discord Rich Presence
pypresence>=4.3.0
